Fix conflicts, missed code and etc
