+++ src/main/java/com/deltabook/controllers/RegistrationController.java
@@ -5,12 +5,13 @@ import com.deltabook.services.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.ModelAttribute;
-import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.*;
 
 @Controller
 public class RegistrationController {
+    private final int WEAK_STRENTH = 1;
+    private final int FEAR_STRENGTH = 5;
+    private final int STRONG_STRENGTH = 7;
 
     @Autowired
     private UserService userService;
@@ -25,6 +26,17 @@ public class RegistrationController {
     public String registerUser(@ModelAttribute User insertedObject) {
         userService.registerUser(insertedObject);
         return "login";
+    }
+    @GetMapping(value = "/checkStrength", produces = {"text/html; charset-UTF-8"})
+    public @ResponseBody
+    String checkString(@RequestParam String password) {
+        if(password.length() >= WEAK_STRENTH & password.length() < FEAR_STRENGTH  )
+            return "слабый";
+        else if(password.length() >= FEAR_STRENGTH & password.length() < STRONG_STRENGTH)
+            return "средний";
+        else if(password.length() >= STRONG_STRENGTH)
+            return "сильный";
+        return "";
     }
 
 }
