+++ src/main/resources/templates/add_user.html
@@ -9,7 +9,7 @@
 <img src="../static/images/logo.png" width="419" th:src="@{images/logo.png}"/>
 <h1>Добавление в друзья</h1>
 <a href="/">Назад в панель пользователя</a>
-<form action="#" th:action="@{/enter_add_friend}" th:object="${user_to}" method="post">
+<form action="#" th:action="@{/enter_add_friend}" th:object="${sendFriendRequest}" method="post">
     <p>Friend NickName:</p>
     <input type="text" th:field="*{friendNickname}" />
     <p>Request Message:</p>
