+++ src/main/resources/templates/login.html
@@ -7,14 +7,16 @@
 </head>
 <body>
 <img src="../static/images/logo.png" width="419" th:src="@{images/logo.png}"/>
-<h1>Вход</h1>
-<form action="#" th:action="@{/login}" method="post">
-    <p>Login:</p>
-    <input type="text" id="login" name="login" autofocus="autofocus"  /></p>
-    <p>Password:</p>
-    <input type="password" id="password" name="password" /></p>
-    <p><input type="submit" value="Submit" /> <input type="reset" value="Reset" /></p>
-</form>
-<a href="/registration">Регистрация</a>
+<div align="center">
+    <h1>Войдите в систему или зарегистрируйтесь</h1>
+    <form action="#" th:action="@{/login}" method="post">
+        <p>Login:</p>
+        <input type="text" id="login" name="login" autofocus="autofocus"  /></p>
+        <p>Password:</p>
+        <input type="password" id="password" name="password" /></p>
+        <p><input type="submit" value="Submit" /> <input type="reset" value="Reset" /></p>
+    </form>
+    <a href="/registration">Регистрация</a>
+</div>
 </body>
 </html>
